w25q128: Support write protect ranges for normal SPI access
This adds the write protect ranges to the W25Q128FV and W25Q128FW
in the "normal" chip definition so it can be used over servo with a
real SPI interface instead of the irritatingly limited hwseq interface.

Also add the set of write protect ranges that are used if the complement
protect bit is set in SR2 which was not accounted for in the original
commit adding ranges to this chip.

Also mark the W25Q128FV chip as tested now that I have been able to
verify read/write/erase/protect functionality via servo.

BUG=chrome-os-partner:37711
BRANCH=none
TEST=test read/write/erase/protect of W25Q128FV over servo

Change-Id: Iadd103ad924d79b80061e958a79bee022e9db71d
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/273551
Reviewed-by: David Hendricks <dhendrix@chromium.org>