dediprog: Use command 0x0B only on older SF100s
As per e-mail with Dediprog, command 0x0B (which is not listed in the
command spec) is used to set voltage level on older Dediprogs. SF100
V6.0.0 and newer as well as all SF600 programmers do not support it.

This patch renames dediprog_device_init() to something more
appropriate  and adds comments for clarity, and only runs it
conditionally if we cannot query the devicestring during init.

Tested with SF100 V:5.1.5 and SF600 V:7.1.4 (with subsequent patches)

BUG=none
BRANCH=none
TEST=Tested on SF100 and SF600 using dwmw2's patch (up next)

Change-Id: I42de7d28401d7ad5be8fcf8a8c165e2614a45960
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293923