Add support for continuous SPI reads
This patch adds support for "unbounded reads" which are for continous
reads to a SPI flash chip instead of reading in page size blocks.

This speeds up the flash process over FTDI by quite a bit, testing
the read of a 16MB part it goes from 2m11s to 0m15s.

BUG=chrome-os-partner:46914
BRANCH=none
TEST=test flash read with and without the patch via servo and in-system
to ensure that the data is read the same in both cases.

Change-Id: I8c5c69f7f7a7295b315885d5090de4a97c54d166
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/308961
Reviewed-by: David Hendricks <dhendrix@chromium.org>