Only verify block erases between writes if we're paranoid
This makes it so that we don't verify that a block was erased before
writing to it unless we are paranoid. For non-paranoid programmers we
will rely on the final verification step only. This greatly reduces
the transaction overhead for external programmers where we can deal
with failures ex post facto.

For veyron_danger (4MB ROM), this reduced the time needed to write
a fully random image to the ROM by about 30 sec (2min 30sec down to
2min) using Servo. No impact using the internal interface.

BUG=none
BRANCH=none
TEST=tested on veyron_danger by writing 4MB of random content with
the "internal" interface and via Servo

Change-Id: Ib0265fe748429210ee220c5d0230dfac4b41d43e
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313016
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>