spi: Allow cached ID bytes to be cleared
This does two things:
1. Removes the static variables from various ID functions and moves
   them to a single struct.
2. Adds a function to clear the struct.

The idea of the original caching mechanism introduced years ago was
simply to speed up probe_flash() by not issuing the same read ID
commands dozens or hundreds of times for a single chip.

That implementation assumed that we would only call probe_flash()
once. However, there are cases when we want to call probe_flash()
multiple times, for example, if using an external programmer and
using different voltages (for probing the chip).

BUG=none
BRANCH=none
TEST=tested with follow-up Dediprog patch for auto-detecting chip

Change-Id: I588bd1b071b76d22fa2c21b2a2ab1e0aec23dcc6
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313752
Reviewed-by: Shawn N <shawnn@chromium.org>