spi25: Add policy-based error handling to spi_read_unbound()
This adds policy-based error handling to spi_read_unbound() so that
it continues past regions where we encounter an error that we intend to
ignore (such as when reading a locked ME region).

Without this, spi_read_unbound() will stop reading when the first error
is encountered and leave the rest of the buffer unfilled.

BUG=none
BRANCH=none
TEST=Tested on Paine with follow-up patches, read operation fills
buffer with 0xff bytes for locked regions and continues to read non-
locked regions so that writes verify successfully.

Change-Id: Id185cba8fb07818c8ceb2c616842b6309d996e1b
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/314237
Reviewed-by: Shawn N <shawnn@chromium.org>