Fix misleading error print in verify_range()
The read buffer offset was being miscalculated in an error print
in verify_range(). This fixes it so that it shows the actual byte
read from the ROM that was mismatched.

BUG=none
BRANCH=none
TEST=Tested with both normal and partial verify (using a layout
file and -i) by comparing with a randomized file making sure
that the values displayed as "Expected" and "Read" make sense.

Signed-off by: David Hendricks <dhendrix@chromium.org>
Change-Id: I2308a31201929b491c2030030db29c1810f0c29c
Reviewed-on: https://chromium-review.googlesource.com/315453
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>