Modify HWSEQ read logic to call ignore_error.
In the current HWSEQ read function implementation of flashrom
When a region is locked or has write-only permissions, read
returns error immediately and the later regions are ignored.

As it is, read happens in 64 byte chunks. If SPI_ACCESS_DENIED
error is returned for a chunk, check if it can be ignored. If
yes, fill the current chunk with 0xff and continue with the
next block instead of returning error.

BUG=chrome-os-partner:49474
TEST=flashrom -p host -r image.bin
When ME is locked, image.bin will contain data of all regions
except ME which will be all 0xFFs.

Change-Id: Ic1eb59d5653820c2535cf7c335d6ea54d21c058a
Signed-off-by: Ramya Vijaykumar <ramya.vijaykumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/322078
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/341117
Tested-by: Chia-Hsiu Chang <chia-hsiu.chang@quantatw.com>
Commit-Queue: Chia-Hsiu Chang <chia-hsiu.chang@quantatw.com>