Change the -E option to allow for partial erase
This simply moves processing of the -E option such that it will run
after old content is read and -i arguments are processed.

The main effect of this is that -E will use the same logic as -w for
determining if a block needs to be erased. This means the chip's
current content will be read first instead of blindly erasing every
block which will likely speed up a full chip erase command. It can
also be used to partially erase a block (surrounding or adjacent
content will be restored).

If the user specifies a filename along with a region, ie when
"-i <region>:<filename>" syntax is used, the content of the
file will be written starting at offset 0 within the region (an error
occurs if the file is too large). This provides a shortcut for users
who currently create binaries to write with 0xff bytes appended at
the end to fill in a region.

BUG=chrome-os-partner:45203
BRANCH=none
TEST=tested on veyron_danger (see below for details)

1. dd if=/dev/urandom of=/tmp/random_4M.bin bs=1M count=4 &&
   dd if=/dev/urandom of=/tmp/random_1M.bin bs=1M count=1 &&
   dd if=/dev/zero bs=1k count=1020 | tr '\000' '\377' > /tmp/ff_1020k.bin
2. echo "200800:3ff7ff region" > /tmp/layout.txt
3. flashrom -p host -w /tmp/random_4M.bin
4. flashrom -p host --ignore-fmap -l /tmp/layout.txt \
   -i region:/tmp/random_1M.bin -E
5. flashrom -p host -r /tmp/cmp.bin
6. cmp /tmp/random_4M.bin /tmp/cmp.bin -i 0:0 -n 0x200800
7. cmp /tmp/random_1M.bin /tmp/cmp.bin -i 0:0x200800 -n 0x100000
9. cmp /tmp/ff_1020.bin /tmp/cmp.bin -i 0:0x300800 -n 1020k
10. cmp /tmp/random_4M.bin /tmp/cmp.bin -i 0x3ff800:0x3ff800

Change-Id: I1c3be38ff6e123bcdf03487fa5088a6baf0bfe04
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/329549
Reviewed-by: Shawn N <shawnn@chromium.org>