cros_ec_dev: Retry commands up to 3 times
If the EC is taking a long time to respond to commands, because perhaps
it has a bug (still being investigated) then EC flash commands may time
out which can be disastrous during dogfood if the EC RO does not write
properly.

This changes the cros_ec_dev driver to retry commands up to 3 times in
case whatever caused a read/erase/write command to fail will succeed if
tried again and not leave the system in a broken state.

The underlying EC bug will hopefully be fixed in time, but this seems
like a good thing to have regardless in case it happens in the future
and so we can update our way out of this current state.

BUG=chrome-os-partner:50586
BRANCH=glados
TEST=tested on a chell system with a "slow EC" behavior where flash
transactions were likely to fail and see that blocks are retried when
they fail and the system flashes successfully.  Also tested manually
on a system that was not showing the "slow EC" symptom by issuing
"waitms 1000" on the EC console while flashing.

Change-Id: I395cf80f511949f89d33c13f0e4c0a20a853dff6
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/329737
Reviewed-by: David Hendricks <dhendrix@chromium.org>