Update GD25Q128 entry
This makes small changes to the existing GD25Q128 entry. The chip
was probably originally added as GD25Q128B, but never really used
in CrOS since there was no write-protect mapping made.

There is a newer version of the chip, the GD25Q128C, which has
quad-IO support and can be distinguished from the GD25Q128B using
SFDP data. Upstream does this, but for now we only care about
the GD25Q128C, so the name is updated to match.

BUG=none
BRANCH=none
TEST=Tested read/write/erase on GD25Q128C.

Change-Id: Ic0046602e20e4b644eaa298288a717878e725e0e
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/332623
Reviewed-by: Shawn N <shawnn@chromium.org>