Verifiy page writes if block was not erased previously
Flashrom is generally expected to bail out immediately if it
encounters an error. erase_and_write_block_helper() currently does
so if it failed to erase a block or if access was denied (ME woes).

There is an overlooked corner case, though, which is if the block was
already erased but cannot be written to due to the flash chip's
write protection.

This fixes it so that we verify writes when we do not do a block
erase operation first. As an optimization, we assume that if a block
erase was done and was successful then we can also write to anywhere
within the block.

There is a small performance hit since we're reading from the flash
where we were skipping earlier. I measured it at <1sec per MB on a
Braswell platform.

BUG=none
BRANCH=none
TEST=tested on Cyan by attempting to write to an empty region with
hardware write-protection enabled, saw flashrom fail immediately
instead of much later during the final verification step.

Signed-off-by: David Hendricks <dhendrix@chromium.org>
Change-Id: I07c40837fa8db2c7e9c802394da47ef39000b209
Reviewed-on: https://chromium-review.googlesource.com/335596
Reviewed-by: Shawn N <shawnn@chromium.org>