Skylake : Correct the flash chip name to "W25Q128.V".
On Skylake platforms like Lars and Chell, flashrom
displays the flash chip name as W25R128FV but the actual chip
is W25Q128FV. Updated this to a generic name with the "."
wildcard "W25Q128.V" to represent one of W25Q128FV, W25Q128JV etc.

Also, the option to probe for a specific chip using --chip was
not working as we only check for the chip name in the list
flashchips, but a new list was also added for platforms that
use HWSEQ host controller interface. This logic was modified
to check for the chip name in both lists.

BUG=chrome-os-partner:51865
BRANCH=none
TEST=flashrom -p host --flash-name should return W25Q128.V on SKL
flashrom -p host --chip W25Q128.V should not return error

Change-Id: I62af688e6b819afa1025a8d367c5fbebcba42837
Signed-off-by: Ramya Vijaykumar <ramya.vijaykumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/336620
Reviewed-by: David Hendricks <dhendrix@chromium.org>