Update list of tested/supported hardware and URLs in print.c
Corresponds to r1471 and r1472 from upstream. A list of the hardware
added/modified as tested and/or supported can be found in the commit
logs from upstream for the given r#'s.
For now, only the additions/modifications have been synced. In a later
revision, the removals may also be synced, but only after it has been
verified that the removal of the specified hardware is also
applciable to chromium flashrom.

BUG=none
BRANCH=none
TEST=compiled successfully, correct output from -L option

Change-Id: I52f42e9fbece2e5ff5c89713a9e3a1264218fb90
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/352246
Reviewed-by: David Hendricks <dhendrix@chromium.org>