Update flashrom to use flashctx struct for flash chip access
Corresponds to r1473 and r1474 from upstream.
For all operations which access hardware, the struct flashctx (flash
context) is now used. Flashctx is similar in structure to flashchip, but
with a few additions. Flashchip is still used for non hardware
operations, such as listing the supported flash chips.

BUG=none
BRANCH=none
TEST=tested correct read/write for auron_paine

Change-Id: I043913d4aa30f47985f96209e2c0a52c98119764
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/353740
Reviewed-by: David Hendricks <dhendrix@chromium.org>