Update all programmer init functions to register bus programmers into common list
Corresponds to r1475 from upstream.
The primary change is that there are no more global variables
to store each programmer type. The flashctx struct now contains
a registered programmer struct, which contains a union of the
three programmer structs (spi, opaque, par). When a programmer
of any type is registered, it is inserted into a list called
registered_programmers.

Note (from upstream):
flashrom -f -c FOO -r forced_read.bin
does not work anymore. We have to find an architecturally clean way
to solve this.

For a full list of changes, please see the corresponding commit
log from upstream.

Important:
This patch is dependent on a subsequent patch which changes cros_ec to
use a global variable. This dependency is indicated below.

BUG=chromium:478356
BRANCH=none
TEST=reads/writes correctly to paine, but needs more testing
CQ-DEPEND=CL:357911

Change-Id: I79199a51b421c27c7122325100cdf2c980ec213d
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/358012
Reviewed-by: David Hendricks <dhendrix@chromium.org>