test_v2: Add support for custom hooks file
This adds an option, --custom-hooks, which allows the user to specify
a script which will be sourced for custom hooks which will be called
before and after each command is run.

The primary use case is to run extra programs needed before flashrom
is used.

BUG=chromium:621715
BRANCH=none
TEST=Used with servo which requires running "dut-control" before and
after a test to enable and disable SPI buffers on its flex cable.

Change-Id: Id5c3829b6b30cebf00dff41e6b3cab0eff6357e2
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362388