Building on test script - endurance/wp modes, logging, code cleanup
Adds endurance and write-protect test types to the test script.
Also, logs of the script output and any failed flashrom runs are
stored in the temporary directory. A future feature might be to
auto-upload these results to gerrit or the flashrom website.

BUG=chromium:621715
BRANCH=none
TEST=needs testing

Change-Id: Iaffee4cf4cfcfcb59209ba7ac17a29a47adf9e92
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/363585
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>