Add ich_descriptor_tool to decode all flash descriptors stored in a flash dump file
Corresponds to r1480 from upstream.
This adds an external utility that shares most of the existing
descriptor decoding source code. This allows access to:
 - the softstraps which are used to configure the chipset by flash content
   without the need for BIOS routines. on ICH8 it is possible to read those
   with FDOC/FDOC too, but this was removed in later chipsets.
 - the ME VSCC (Vendor Specific Component Capabilities) table. simply put,
   this is an SPI chip database used to figure out the flash's capabilities.
 - the MAC address stored in the GbE image.

Parts of this patch already existed in the chromium flashrom code, but in
order to prevent future merge conflicts, I took a "use theirs" strategy for
merging this upstream patch.

BUG=chromium:478356
BRANCH=none
TEST=needs testing

Change-Id: I5ede2c466cf6cda7c4ee20863e188d856aeb1ce3
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/365861
Reviewed-by: David Hendricks <dhendrix@chromium.org>