layout.c: change name and return type of find_next_included_romentry
Corresponds to r1481 from upstream.
The function, now called get_next_included_romentry, returns
a pointer to a rom_entry instead of the index of that rom_entry.
This prevents the calling functions from having to access the data
structure holding the entries multiple times subsequently. This change
is mainly for improving data separation and code readability.

BUG=chromium:478356
BRANCH=none
TEST=needs testing

Change-Id: I90cd61208b0d28ed7edfb350f70ccb03aa54c900
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/365892
Reviewed-by: David Hendricks <dhendrix@chromium.org>