Reorder list of options to test
Corresponds to r1485 from upstream.
Commit logs from upstream:
If a chip is unknown the user is asked to test and report the result to
the mailing list. Having `-VE` listed as the last option can result in
an unbootable system for users not knowing what the command does, since
rebooting the system after that command is fatal since the flash chip is
empty.

BUG=chromium:478356
BRANCH=none
TEST=none

Change-Id: I43d9dd63f2c1cc58f2e98f19b290eb044e7c8890
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/367102
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>