Patch up reverted merges and create global cros_ec variable
This patch essentially reapplies CL:357911, which was reverted
along with 3 other patches, so that future patches can be rebased
on top of this. The reverted patches were intended to combine the
global spi/par/opaque programmer variables into a union as a field
in a new struct, registered_programmer (this is implemented in
upstream). However, because of some complications with certain
cases involving multiple programmers, we are keeping the globals.

BUG=chromium:635405
BRANCH=none
TEST=successfully reads on Lumpy, Jaq, and Link

Change-Id: Ife8f24d852ba0e6dba6051d05837789ff8600714
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/368480
Reviewed-by: David Hendricks <dhendrix@chromium.org>