Warnings for one-time programmable (OTP) memory
Corresponds to r1493 from upstream.
Commit log from upstream:
Some flash chips contain OTP memory that we cannot read or write (yet). This
prohibits us from cloning them, hence warn the user if we detect it. Not all
variations of the tagged chips contain OTP memory. They are often only
enabled on request or have there own ordering numbers. There is usually no
way to distinguish them. Because this is a supposedly seldomly used feature
the warning is shown in with dbg verbosity.

The manpage is extended to describe the backgrounds a bit.

This patch is based on the idea and code of Daniel Lenski.

BUG=chromium:478356
BRANCH=none
TEST=needs testing

Change-Id: Id6e7d1b8409e02b59aa451448490badd09e25200
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/368880
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/428271
Tested-by: Jongpil Jung <jongpil19.jung@samsung.com>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Jongpil Jung <jongpil19.jung@samsung.com>