test_v2: add wp hooks files for servo-v2
CL:363585 added the functionality for wp-hooks, or
the functions that an external programmer can use
to enable/disable write protection and turn the
current wp setting on/off. This is an example file
for how the wp hooks might look. This file corresponds
to the servo-v2 commands for running on veyron_jaq.

BUG=chromium:621715
BRANCH=none
TEST=none

Change-Id: I2002ee0966eb7f333dab0826e417daee50adeff0
Signed-off-by: Souvik Ghosh <souvikghosh@google.com>
Reviewed-on: https://chromium-review.googlesource.com/368882
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>