Fix conflicting FEATURE_OTP and FEATUTRE_ERASE_TO_ZERO #defines
FEATURE_OTP and FEATURE_ERASE_TO_ZERO were introduced in different
branches at different times and ended up conflicting. This changes
FEATURE_ERASE_TO_ZERO to (1 << 9).

While we're at it, condense the list a bit so it's not as easy to
overlook next time we're merging a bunch of patches...

BUG=chromium:642243
BRANCH=none
TEST=flashing didn't fail when FEATURE_OTP added (tested using
W25Q128FW on Reef)

Change-Id: Idd06f1c582c35f9d6b696abaabf4c6f38e7f5858
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/377693
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>