ichspi: Don't override flash chip test status
Although hardware sequencing blurs the responsibility of the chip and
PCH, the test status bits should only be set in the chip's definition.

This has the side-effect of getting rid of an annoying warning due
to unbounded read test status not being set by the ICH/PCH drivers.

BUG=none
BRANCH=none
TEST=tested on reef, no more warning about TEST_OK_MASK

Change-Id: I5d7fd4ad2c24a6cbbf31c4f646e798e2d777879e
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/385186
Reviewed-by: Duncan Laurie <dlaurie@google.com>