Make extract_param() not change the input buffer
extract_param() was introduced many years ago in upstream and has
always removed the "needle" from the "haystack".

The best reason for that approach is to detect if there are unused
parameters specified which could be an indication that the user
misspelled a parameter or some other silly mistake.

More importantly though, multiple programmers init functions should
be able to parse the same parameters so that if init for one device
fails, we can gracefully fall thru to try another which may use the
same parameters. This is especially true for the ECs we support.

Long-term, the fix should be to parse the programmer parameters
into a linked list so that we can detect unused parameters, detect
duplicate parameters, and possibly other useful things. In the short-
term though we need to make extract_param() non-destructive.

BUG=chromium:658207
BRANCH=none
TEST=tested on reef with follow-up patch

Change-Id: I824be388cc3b094c889683f71485a8af6b760006
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/402057
Reviewed-by: Hung-Te Lin <hungte@chromium.org>