cros_ec_dev: Demote error print for command failure
Currently if a command fails (not due to communication) cros_ec_dev
will print an error message to the console. This should be a debug
print, and the caller should print an error if it's really fatal.
For example, EC_CMD_FLASH_SPI_INFO is expected to fail when the EC
doesn't use external SPI flash.

(h/t gwendal@chromium.org for pointing this out)

BUG=none
BRANCH=none
TEST=needs testing

Change-Id: Ia73fba441259c0221263d120b932e9dba3e9693f
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/406732