test_v2: Strip basename from $LAYOUT_FILE when grepping for region name
If a layout file is used, it is copied to $LOCAL_TMPDIR. This file
can then be grepped to search for the specified region.

But $LAYOUT_FILE as specified by the user is an absolute path which
may be on a remote system. So we need to strip the path when we're
looking at the file stored in $LOCAL_TMPDIR.

Opportunistically update the comment while we're at it.

BUG=chrome-os-partner:59356
BRANCH=none
TEST=did partial clobber tests on Kefka

Signed-off-by: David Hendricks <dhendrix@chromium.org>

Change-Id: Ia09613deead6430a2dbb47e2558c70ce9de09e31
Reviewed-on: https://chromium-review.googlesource.com/407441
Commit-Ready: David Hendricks <dhendrix@chromium.org>
Tested-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: David Hendricks <dhendrix@chromium.org>
Reviewed-by: YH Lin <yueherngl@chromium.org>