test_v2: Fix print_and_log uses for old printf users
print_and_log replaced usage of printf in most cases. However, it
can only take only 1 parameter and there were still a few parts of
the script that used printf semantics where more than 1 parameter
was used, for example, to calculate a value to be formatted by printf.

This fixes the few cases where printf semantics are used so that the
calculation is performed beforehand and included in the string passed
to print_and_log.

BUG=chrome-os-partner:59356
BRANCH=none
TEST=Did a partial clobber test on Kefka, prints turned out correctly

Change-Id: Id34b2d95cdc1deb6de3548b45bf47953efa388c7
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/407444
Reviewed-by: YH Lin <yueherngl@chromium.org>