test_v2: Add -v/--voltage arg and make Servo hooks generic
This allows the user to pass a voltage value into the test script
which will then get passed into the Servo hooks.

Although it adds slightly more complexity to the command-line, it
actually simplifies things from a scripting perspective since the user
doesn't need to select a different set of hooks depending on the
voltage of the chip. They just need to pass in the same voltage which
they use to enable the SPI buffers via dut-control.

BUG=chromium:621715
BRANCH=none
TEST=tested with follow-up patches

Change-Id: Iac34dd89c7241a4171818a81126088fbdd01f4cf
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/410147
Reviewed-by: Martin Roth <martinroth@chromium.org>