test_v2: Add custom_hook_sanity_check()
This adds a sanity check for when we're using custom hooks. The
first use case is to check that dut-control can obtain certain values
that are set by other some other hooks.

The hook is intended to run early so that we don't get confusing
errors several seconds after the script begins.

BUG=chromium:621715
BRANCH=none
TEST=ran with servod listening on a non-default port, script
erred out sooner rather than later.

Change-Id: I63d83f6d6fc9e26caadfb4c05c73e07f4018c0f9
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/422997
Reviewed-by: Martin Roth <martinroth@chromium.org>