ichspi: Fill in correct chip details when HWSEQ is used.
Currently the vendor and chip vendor and name are hardcoded to
Winbond W25Q128.V whenever hardware sequencing is used. This is an
artifact of an earlier implementation.

This patch attempts to fill in proper chip details by issuing an RDID
command to the chip and matching the manufacturer/model against the
main flashchips[] array. If found, we'll simply copy details from the
chip found.

This was reverted in commit 807acd7d, but with new follow-on patches
should be good to submit again.

BUG=chrome-os-partner:59223
BRANCH=none
TEST=tested on Reef

Original-Change-Id: I99e264ee93b53d5aea73aad30d95beb084e4f0a3
Original-Reviewed-on: https://chromium-review.googlesource.com/423178
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Change-Id: I490f2ed786be09866f82829eff4f41e571a9c548
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/448164
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>