ichspi: Fill in correct chip details when HWSEQ is used.
Currently the vendor and chip vendor and name are hardcoded to
Winbond W25Q128.V whenever hardware sequencing is used. This is an
artifact of an earlier implementation.

This patch attempts to fill in proper chip details by issuing an RDID
command to the chip and matching the manufacturer/model against the
main flashchips[] array. If found, we'll simply copy details from the
chip found.

BUG=chrome-os-partner:59223
BRANCH=none
TEST=tested on Reef

Change-Id: I99e264ee93b53d5aea73aad30d95beb084e4f0a3
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/423178
Reviewed-by: Aaron Durbin <adurbin@chromium.org>