flashchips: Update hardware sequencing and W25Q128 entries
The Winbond W25Q128 series chips and Intel hardware sequencing have
a bit of an interesting/confusing history. This patch attempts to
straighten things out.

Some background:
- Support for some W25Q128 parts imported from upstream a long time ago.
- W25Q128FW was added for Smaug. W25Q128FW is a 1.8V variant which
shares model ID with some uncommon 3.3V variants of the W25Q128
(including the W25Q128FV operating in QPI mode).
- To further confuse things, when hardware sequencing for Apollo
Lake (Reef) was added, we hardcoded "W25Q128.V" as the name of
the chip. This was intended to account for some variants, but the
platform which uses this method (Reef) actually has a W25Q128FW
like Smaug.

This patch makes the cosmetic change to the original W25Q128 entry
so that it shows "W25Q128.V" as the name when it's found, indicating
a possible match for W25Q128BV, W25Q128FV (SPI mode), or W25Q128JV.

The follow-up patch will make it so that platforms which use hardware
sequencing (e.g. Reef) will correctly identify the chip. For Reef, this
will result in "flashrom -p host --flash-name" returning "W25Q128FW"
instead of "W25Q128.V".

This patch opportunistically adds missing attributes to the W25Q128FW
entry, namely the OTP feature bit and voltage. The voltage attribute
is set to 1.8V for now. It's a bit unfortunate since that might
make things difficult for users of the other variants which share the
same model ID. They may need to specify voltage on the command-line
or create a duplicate entry with a different name that can be
specified using "-c" on the command-line.

This was reverted in commit 5a924b10, but with new follow-on patches
should be good to submit again.

BUG=chrome-os-partner:57193
BRANCH=none
TEST=tested with follow-up patch

Change-Id: I01e98f447afe61c6d9d6d6143f494a530bca070a
Original-Change-Id: Id1913eb7377ca5c6aadced4e148d2d20b813a013
Original-Reviewed-on: https://chromium-review.googlesource.com/423177
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Signed-off-by: David Hendricks <dhendrix@chromium.org>
Signed-off-by: Martin Roth <martinroth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/448163
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/489474
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>